Update and rename Assignment1.2.cpp to BasicPrintStatementsUnFormatted.cpp
You are trying to make ASCII art. Write a program that produces the following output:

CCCCCCCCC       ++              ++
CC              ++              ++
CC        ++++++++++++++  +++++++++++++++
CC        ++++++++++++++  +++++++++++++++
CC              ++              ++
CCCCCCCCC       ++              ++
Note: The letter C in the output must be uppercase. There is one column of space between the "C" and the first plus symbol and two columns of space between the two plus symbols.